refactor(DayManager): DayManager -> DateInfoProvider로 클래스 이름 변경
DayManager의 역할을 보았을때, 매니저 정도의 책임을 가지지 않아서, service 패키지로 분류할 예정입니다.
